ai blog

+++ blog/app/Http/Controllers/ArticleController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Article;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -52,6 +54,7 @@ class ArticleController extends Controller
         $article->title = $request->title;
         $article->body = $request->body;
         $article->category_id = $request->category_id;
+        $article->user_id = Auth::id();
         $article->save();
 
         return redirect('/articles');
@@ -60,8 +63,12 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
 
-        return redirect("/articles")->with("info", "Deleted an article");
+        if(Gate::allows("delete-article", $article)) {
+            $article->delete();
+            return redirect("/articles")->with("info", "Deleted an article");
+        }
+
+        return back()->with("info", "Unauthorized to delete");
     }
 }
